refactor(elaborate): carve param/override/const-eval into elaborate::params — move-only (P4 phase 2a)
Converts src/elaborate.rs into a directory module (src/elaborate/mod.rs)
and extracts the param-resolution family into src/elaborate/params.rs:
override collection (collect_raw_overrides_from_body, record_inst),
variant discovery (compute_all_variants, find_varying_params,
make_variant_name), per-variant elaboration (elaborate_module_variant,
rewrite_inst), derived-param recompute (recompute_derived_params,
expr_references_params), default computation (compute_defaults_with_enums),
and elaborate-side const-eval (try_eval_i64, try_eval_bool) — the same
try_eval_i64 that typecheck.rs's where-clause enforcement (PR #695) calls
via crate::elaborate::try_eval_i64.

Move-only: 5 functions bumped from private to pub(crate) purely for
cross-module access (matching the sim_codegen split's pub(crate)
convention); try_eval_i64 stays pub fn and is re-exported from mod.rs so
every existing call site (internal bare calls + typecheck.rs's qualified
calls) resolves unchanged. No logic, formatting, or diagnostic text
changed within any moved function body.
